pyproject.toml
[project]
name = "es-knowledge-base-mcp"
version = "0.1.0"
description = "An MCP server for crawling documentation, writing it to Elasticsearch for vector search retrieval by an LLM."
authors = [
    {name = "Bill Easton",email = "williamseaston@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "elasticsearch (>=8.17.2,<9.0.0)",
    "pyyaml>=6.0.2,<7.0.0",
    "anyio (>=4.9.0,<5.0.0)",
    "fastmcp>=2.1.2",
    "aiohttp>=3.11.14",
    "pydantic-settings>=2.3.4,<3.0.0",
    "aiodocker>=0.24.0",
    "pydantic>=2.0.0,<3.0.0",
    "asyncio (>=3.4.3,<4.0.0)",
    "requests>=2.32.3,<3.0.0",
    "beautifulsoup4>=4.12.3,<5.0.0",
    "async-lru (>=2.0.5,<3.0.0)",
    "markdownify>=1.1.0,<2.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0","pip"]
build-backend = "poetry.core.masonry.api" 

[project.scripts]
es-knowledge-base-mcp = "es_knowledge_base_mcp.server:run"

[project.optional-dependencies]
dev = [
    "ruff>=0.11",
    "pytest>=8.3,<9.0.0",
    "pytest-asyncio>=0.26,<0.27.0",
    "syrupy>=4.9",
    "requests-mock>=1.12,<2.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.poetry]
packages = [
    { include = "es_knowledge_base_mcp", from = "src" },
]
[tool.ruff]
target-version = "py312"
respect-gitignore = true
lint.fixable = ["ALL"]
lint.ignore = [
    "COM812",
    "PLR0913", # Too many arguments, MCP Servers have a lot of arguments, OKAY?!
]
lint.extend-select = [
    "A",
    "ARG",
    "B",
    "C4",
    "COM",
    "DTZ",
    "E",
    "EM",
    "F",
    "FURB",
    "I",
    "LOG",
    "N",
    "PERF",
    "PIE",
    "PLR",
    "PLW",
    "PT",
    "PTH",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "TC",
    "TID",
    "TRY",
    "UP",
    "W",
]

line-length = 140

[tool.ruff.lint.extend-per-file-ignores]
"**/tests/*.py" = [
    "S101", # Ignore asserts
    "DTZ005", # Ignore datetime.UTC
    "PLR2004", # Ignore magic values
]
"**/references/*" = ["ALL"]
"template/*" = ["ALL"]


[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["es_knowledge_base_mcp"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
